+++ flutter_inappwebview_ios/ios/flutter_inappwebview_ios/Sources/flutter_inappwebview_ios/HeadlessInAppWebView/HeadlessInAppWebView.swift
@@ -38,7 +38,7 @@ public class HeadlessInAppWebView: Disposable {
             } else {
                 view.frame = CGRect(x: 0.0, y: 0.0, width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height)
             }
-            if let keyWindow = UIApplication.shared.keyWindow {
+            if let keyWindow = UIApplication.shared.keyWindowCompat {
                 /// Note: The WKWebView behaves very unreliable when rendering offscreen
                 /// on a device. This is especially true with JavaScript, which simply
                 /// won't be executed sometimes.