+++ flutter_inappwebview_ios/ios/flutter_inappwebview_ios/Sources/flutter_inappwebview_ios/UIApplication/VisibleViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 extension UIApplication {
 
     var visibleViewController: UIViewController? {
-        guard let rootViewController = keyWindow?.rootViewController else {
+        guard let rootViewController = keyWindowCompat?.rootViewController else {
             return nil
         }
         return getVisibleViewController(rootViewController)